modals kinda working but not really :(

+++ routes/web.php
@@ -17,7 +17,7 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [CategoryController::class ,'index']);
+Route::get('/', [CategoryController::class ,'index'])->name('welcome');
 Route::fallback(function () {
     return response()->view('errors', [], 404);
 });
@@ -35,18 +35,18 @@ Route::get('/signup',function(){
 Route::post('/signup', [UserController::class,'register'])->name('register.user');
 Route::get('/admin',function(){
     return view('Layouts.Admin.dashboard');
-})->middleware('auth')->name('admin.dashboard');
+})->middleware(['auth','admin'])->name('admin.dashboard');
 Route::get('/logout',[UserController::class,'logout']);
 
-Route::get('/all-products',[ProductController::class,'getProducts'])->middleware('auth')->name('product.all');
-Route::get('/add-product',[ProductController::class,'getCategories'])->middleware('auth');
-Route::post('/add-product',[ProductController::class,'add'])->middleware('auth')->name('product.add');
-Route::delete('/delete-product/{id}',[ProductController::class,'destroy'])->middleware('auth')->name('product.delete');
+Route::get('/all-products',[ProductController::class,'getProducts'])->middleware(['auth','admin'])->name('product.all');
+Route::get('/add-product',[ProductController::class,'getCategories'])->middleware(['auth','admin']);
+Route::post('/add-product',[ProductController::class,'add'])->middleware(['auth','admin'])->name('product.add');
+Route::delete('/delete-product/{id}',[ProductController::class,'destroy'])->middleware(['auth','admin'])->name('product.delete');
 
 Route::get('/add-category',function(){
     return view('Layouts.Admin.addCategory');
-})->middleware('auth');
-Route::get('/all-categories',[CategoryController::class,'getCategories'])->middleware('auth')->name('category.all');
-Route::post('/add-category',[CategoryController::class,'add'])->middleware('auth')->name('category.add');
-Route::delete('/delete-category/{id}',[CategoryController::class,'destroy'])->middleware('auth')->name('category.delete');
+})->middleware(['auth','admin']);
+Route::get('/all-categories',[CategoryController::class,'getCategories'])->middleware(['auth','admin'])->name('category.all');
+Route::post('/add-category',[CategoryController::class,'add'])->middleware(['auth','admin'])->name('category.add');
+Route::delete('/delete-category/{id}',[CategoryController::class,'destroy'])->middleware(['auth','admin'])->name('category.delete');
